refactor<view> : io,validate 패키지 생성후 클래스 이동
