+++ src/main/kotlin/baseball/view/validate/InputValidate.kt
@@ -1,4 +1,4 @@
-package baseball.view
+package baseball.view.validate
 
 import baseball.ErrorMessage
