+++ src/test/kotlin/baseball/InputValidateTest.kt
@@ -1,6 +1,6 @@
 package baseball
 
-import baseball.view.InputValidate
+import baseball.view.validate.InputValidate
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
